Crate massa_cipher

MASSA Cipher crate

massa-cipher uses AES-GCM

AES-GCM is a state-of-the-art high-performance Authenticated Encryption with Associated Data (AEAD) that provides confidentiality and authenticity.

To hash the password before using it as a cipher key, we use the PBKDF2 key derivation function as specified in RFC 2898.

The AES-GCM crate we use has received one security audit by NCC Group, with no significant findings.

Enums

CipherError

Cipher error

Functions

decrypt

Decryption function using AES-GCM cipher.

encrypt

Encryption function using AES-GCM cipher.